+++ core/__init__.py
@@ -5,3 +5,4 @@ import poisoners
 import responder
 import payloads
 import autocrack
+import services
