+++ core/servers/redirect_server.py
@@ -1,14 +1,12 @@
-from BaseHTTPServer import BaseHTTPRequestHandler, HTTPServer
-from SocketServer import ThreadingMixIn
-from multiprocessing import Process
 import time
 import sys
 import random
 import string
 
-# lol, i mean it works...
-bind_addr = None
-bind_port = None
+from BaseHTTPServer import BaseHTTPRequestHandler, HTTPServer
+from SocketServer import ThreadingMixIn
+from multiprocessing import Process
+
 upper_alnum = string.ascii_uppercase + string.digits
 
 class RedirectHandler(BaseHTTPRequestHandler):
@@ -17,7 +15,7 @@ class RedirectHandler(BaseHTTPRequestHandler):
 
         s.send_response(302)
         share_path = ''.join(random.choice(upper_alnum) for _ in range(8))
-        new_location = 'file://%s/%s' % (bind_addr, share_path)
+        new_location = 'file://10.0.0.1/%s' % (share_path)
         s.send_header('Location', new_location)
         s.end_headers()
 
@@ -36,16 +34,6 @@ class RedirectHandler(BaseHTTPRequestHandler):
 class ThreadedHTTPServer(ThreadingMixIn, HTTPServer):
     ''' yay '''
 
-def run_redirect_server(bind_addr, bind_port):
-
-    server_class = ThreadedHTTPServer
-    httpd = server_class((bind_addr, bind_port), RedirectHandler)
-    try:
-        httpd.serve_forever()
-    except KeyboardInterrupt:
-        pass
-    httpd.server_close()
-
 class RedirectServer(object):
 
     instance = None
@@ -56,27 +44,26 @@ class RedirectServer(object):
             instance = RedirectServer()
         return instance
     
-    def configure(self, b_addr, b_port=80):
-        
-        global bind_addr
-        global bind_port
+    def configure(self, bind_addr, bind_port=80):
 
-        # idfk
-        bind_addr = b_addr
-        bind_port = b_port
+        self.bind_addr = bind_addr
+        self.bind_port = bind_port
 
     @staticmethod
-    def _start(args):
+    def _start(bind_addr, bind_port):
 
-        run_redirect_server(args['bind_addr'], args['bind_port'])
+        server_class = ThreadedHTTPServer
+        httpd = server_class((bind_addr, bind_port), RedirectHandler)
+        try:
+            httpd.serve_forever()
+        except KeyboardInterrupt:
+            pass
+        httpd.server_close()
 
     def start(self):
     
-        args = {
-            'bind_addr' : bind_addr,
-            'bind_port' : bind_port,
-        }
-        self.proc = Process(target=self._start, args=(args,))
+        args = (self.bind_addr, self.bind_port,)
+        self.proc = Process(target=self._start, args=args)
         self.proc.daemon = True
         self.proc.start()
         time.sleep(4)
@@ -85,4 +72,3 @@ class RedirectServer(object):
 
         self.proc.terminate()
         self.proc.join()
-
